+++ backend/database/migrations/2023_05_29_073717_create_order_detail_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
             $table->id();
             $table->bigInteger('order_id')->unsigned();
             $table->bigInteger('product_id')->unsigned();
-            $table->float('unit_price')->nullable();
+            $table->decimal('unit_price',12,2)->nullable();
             $table->bigInteger('quantity')->nullable();
-            $table->float('promotion_price')->nullable();
+            $table->decimal('promotion_price',12,2)->nullable();
             $table->bigInteger('create_user')->nullable();
             $table->bigInteger('modified_user')->nullable();
             $table->bigInteger('state')->default(1)->comment('1:live 9:kill');
